Fix the move constructor of Option.
Some bit of background on this. It looks like (the class is Foo):

  Foo(Foo&&) noexcept = default;

Is not portable as it won’t compile on clang as of today. The single
solution I found was to explicitly implement that constructor
(Option<T>). I’m not sure whether it’s linked to the fact the class is
templated or what.