Fix a bug in live with the AESCipher (that doesn’t like empty passwords).
This fixes the first part of the problem. I’ll open a Jira ticket to fix
AES/PBKDF2 when the key / passphrase is empty.

Also, note: we might want to change the interface of the WalletPool
creation. This would be a breaking change, though. I’ll also open a Jira
ticket for that.