Fix android build on MacOSX. [skip ci]
I guess the include/darwin stuff is fixed on some implementation (it’s
not even linked to Mojave vs. not Mojave). Looks like that it will
depend on which implementation of JNI you have. In my case, I have the
default one, which has that darwin include dir that needs to be added.

I feel like my soul is gone and my sanity evaporated.