Review password encoding.
This commit removes and replaces the optional string encoding by a
mandatory string. It means that both Optional::NONE and Optional("") are
mapped to "": empty passwords are never accepted for encoding, so they
no mean “no password”.

As soon as this commit gets merged, the bindings must be updated
accordingly as well as any dependent application as this is a breaking
change.