Changed the ownership mechanisme of LevelDB in PreferencesBackend.
Instead of having a non-owning unordered_map and lots of owning
shared_ptr everywhere in the codebase, we now have an owning table and
weak_ptr everywhere. The idea is that, if get a
shared_ptr<ledger::core::PreferencesBackend>, the underlying database is
handled via a weak_ptr. Which means that you have to lock it in order to
use the database, which allows us to destroy the database and observe it
everywhere.

Fixes LLC-198