Add configuration for RIPPLE_LAST_LEDGER_SEQUENCE_OFFSET.
This PR supplements #290 as some teams (i.e. Vault, ping @onyb) requires
an higher value for the offset (default is 4 which is too short to get
all the signatures).

Possible alternative was to set maxLedgerVersion to `null` [but this is
highly
discouraged](https://xrpl.org/reliable-transaction-submission.html#lastledgersequence).

Relates to LLC-346.