Add WalletDatabaseHelper and remove the concept of “pool names”.
Pool names are basically a namespacing feature of libcore I have no idea
what we would use it for. Currently, both Live and Vault don’t use it
(they set it to a random value, or juste “Ledger Live” for instance).

However, “pool names” were a good idea misnamed. We want them to be
called “tenants” instead. A tenant is just a way to implement
namespacing by client applications.