Add implementation of getDryrunGasLimit
- Enabling modification of "amplifier", "from" and "data" fields in the
underlying explorer API call means the interface of the explorer needs
to change.

- Separate GET /estimate and POST /est. endpoints
Behaviour is too different to be crammed in the same function. Also,
this ensures forward compatibility.

- Add EthereumGasLimitRequest type
This structure encapsulates the requests passed to the backend

- Catch and throw 5** errors for ill-formed requests to the POST
underlying endpoint.
These errors were handled cleanly in the GET request, giving a default
gas_limit value.
This case isn't handled anymore when using the POST endpoint, so for the
time being all errors are "recovered" into throwing an INVALID_ARGUMENT
exception instead.