Fix OperationQuery in ledger-core to authorize coin to override performExecute.
This fix is a bit tricky but is necessary with the current definition of
OperationQuery. The main problem is due to the fact that performExecute
introduces some columns into scope (among date). Even though they’re
namespaced (via a `o.`), the QueryBuilder can inject filters based on
those columns, which are not namespaced here. That causes ambiguities if
performExecute is overriden with other tables which also have a date
column (but it can be any other column).