Disable dropping the ledger-core database.
This fix is required to prevent the freshResetAll from becoming crazy.
The current problem with this is due to the fact that coins don’t
currently register themselves as using the database. So when we drop the
database, we forget to drop the coin tables.

A fix is required to do this. See the Jira ticket LLC-509 for follow-up.