Add SSL_SUPPORT configuration to build Core library with SSL support.
Before we didn't need SSL support because we were using only Crypto libraries on OpenSSL.
Eventually we were using SSL for tests because we were implementing our own Http client with SSL support.
We didn't face that issue before with any client because Http clients are provided by the host of the Core library.

We are not only building with support of SSL but in this commit we initialize of course the SSL libraries when using PG backend.