Add support for failover behavior in soci connection pools.
With the current version of soci, when a `connection_pool` is created, a
buffer is allocated with, for each element in the buffer, a pair `(bool,
session*)`. The boolean value indicates whether the session is free or
not (this is pretty bad way to encode a pool, but whatever). When a
session is freed, the boolean gets changed and then opening a new
session will reuse that slot.

However, when a session has been idleing for a while, it’s possible that
the database backend (e.g. PostgreSQL) kills the connection held in the
session for timeout. The `soci` library doesn’t see that and assumes the
session is still valid -> we hit errors, which get propagated as
`std::runtime_error` in our codebase.

This commit fixes the problem by introducing a `is_alive()` function on
the session that allows a database backend to execute some logic to
check the status of the session before using it. SQLite3 doesn’t drop
sessions (no timeout) but PostgreSQL does), so we can implement that
logic in the PostgreSQL backend.

It should be fix LVS-92.