Final fix for LVS-92 (PostgreSQL timeout deconnection).
The PQexec command seems to be needed to actually witness the drop of
connection (i.e. socket byte error). Then PQreset allows us to
gracefully restore the connection.

I know the interface is not perfect but that’s completely due to soci
lacking that feature in this version. Once we upgrade, we should have
that failover behavior by default, so this current patch will be
completely dropped.