LLC-484 – Turn some 64-bit integers to 32-bit integers in the interface.
Most of those types are used for stuff like query limits, etc. which are
all far from ever reaching the maximum value for a 32-bit integers, and,
also, we use 32-bit a lot in the rest of the codebase.

This commit is for LLC-484 where, on ReactNative, 64-bit precision
integers are not a possible thing — currently, the Ledger Live team has
an internal patch command they run to automatically turn the `long`
argument into `int` argument.

Once this commit is merged, we need to merge the ReactNative PR that
will update the interface so that they can use it.