LLC-643 – Fix negative / non-matching balances on Tezos (originated accounts).
The problem came from the Babylon 2.0 protocol update, which introduced
a new way to spend funds and pay fees. The code is pretty
self-explanatory, so feel free to read the content of the commit for
further information, but basically, a mirror fix is needed on the
implicit accounts, as they now have to spend the fees of originated
accounts as well.