AlgorandAddress
Implementation of AlgorandAddress, based on temporary (incorrect) implementation of SHA-512/256 wrapper.

I also put the wrapper part on a separate clean branch based on design/modularization, to send a PR to libcore when it's ready. I copy the modifications here so that we can keep moving while it's not merged in libcore yet.