fix: weird crash in wd integration tests when loading libcore.
The fix is to avoid using std::regex constructor of global static variables
using lazy initialization of static variables. The crash happens when the
WD tries to load the libcore and static regex variable are being
initialized. It looks like the locale is not yet correctly defined and the
regex constructor does not accept this case.